+++ pystregy/model/quote.py
@@ -1,11 +1,10 @@
 from datetime import datetime
-from decimal import Decimal
 
 class Quote():
     __slots__ = 'time', 'open', 'high', 'low', 'close'
 
     def __init__(self, time: datetime, 
-        open: Decimal, high: Decimal, low: Decimal, close: Decimal
+        open: float, high: float, low: float, close: float
     ):
         self.time = time
         self.open = open
